fix: remove unnecessary cast in thirds score (analyzer clean)
The Task 4 review ran flutter test but not dart analyze, so an
unnecessary_cast lint slipped through. Use 1.0 literal so score is
double without a cast.

+++ lib/analysis/thirds.dart
@@ -31,7 +31,7 @@ ThirdsAlignment computeThirds(double cx, double cy,
       }
     }
   }
-  final score = 1 - (math.min(bestD / 0.4, 1.0) as double);
+  final score = 1.0 - math.min(bestD / 0.4, 1.0);
   final dx = bestX - cx;
   final dy = bestY - cy;
   final parts = <String>[];
